refactor: Refactored RecognitionDataset
