+++ doctr/datasets/__init__.py
@@ -1,8 +1,8 @@
 from .utils import *
 from .vocabs import *
-from .loaders import *
 from .funsd import *
 from .sroie import *
 from .cord import *
 from .loader import *
 from .detection import *
+from .recognition import *
