--- doctr/datasets/loaders/__init__.py
@@ -1,1 +1,0 @@
-from .recognition import *
